Separates the functionality for readability
